+++ packages/authentication/modules/user/useJWTUser/types.ts
@@ -1,5 +1,3 @@
-import type { ServerSideRenderingOption } from '@baseapp-frontend/utils'
-
 import UserApi from '../../../services/user'
 import type { CustomJWTKeyNames } from '../../../types/auth'
 import type { CustomUseQueryOptions } from '../../../types/react-query'
@@ -8,7 +6,6 @@ type ApiClass = Pick<typeof UserApi, 'getUser'>
 
 export interface UseJWTUserOptions<TUser>
   extends CustomJWTKeyNames,
-    ServerSideRenderingOption,
     CustomUseQueryOptions<TUser, unknown, TUser> {
   ApiClass?: ApiClass
 }