+++ packages/design-system/components/web/dropzones/Dropzone/types.ts
@@ -1,4 +1,4 @@
-import { CSSProperties, FC, HTMLAttributes, InputHTMLAttributes } from 'react'
+import { CSSProperties, FC, HTMLAttributes, InputHTMLAttributes, ReactElement } from 'react'
 
 import { SvgIconProps } from '@mui/material'
 import { Theme } from '@mui/material/styles'
@@ -23,7 +23,7 @@ export interface BaseDropzoneProps {
   maxFileSize?: number
   DropzoneOptions?: Partial<DropzoneOptions>
   InputProps?: InputHTMLAttributes<HTMLInputElement>
-  title?: string | JSX.Element
+  title?: string | ReactElement
   includeActionButton: boolean
   InputContainerStyle?: CSSProperties
   multiple?: boolean