+++ packages/utils/functions/language/getLanguage/index.ts
@@ -1,8 +1,12 @@
-import { LANGUAGE_COOKIE_NAME } from '../../../constants/cookie'
-import type { ServerSideRenderingOption } from '../../../types/server'
-import { getCookie } from '../../cookie'
+import { getItem } from 'expo-secure-store'
 
-export const getLanguage = (
-  accessKeyName = LANGUAGE_COOKIE_NAME,
-  { noSSR = false }: ServerSideRenderingOption = {},
-) => getCookie<string>(accessKeyName, { noSSR })
+import { LANGUAGE_COOKIE_NAME } from '../../../constants/cookie'
+import { getCookie } from '../../cookie'
+import { isMobilePlatform } from '../../os'
+
+export const getLanguage = (languageKey = LANGUAGE_COOKIE_NAME) => {
+  if (isMobilePlatform()) {
+    return getItem(languageKey)
+  }
+  return getCookie<string>(languageKey)
+}